Move ModifyConfig (and some related functions) into the clientcmd pkg.
